+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 ## v49.x / v50 (In progress)
-- PC sprites: Added Exit Mouse, Duster, Save Frog, Edgar, Rek, Dorothy, and Sophia.
+- PC sprites: Added Exit Mouse, Rope Snake, Duster, Flint, Save Frog, Edgar, Rek, General Leo, Prince of Cannock, Prince of Midenhall, Princess of Moonbrooke, Dorothy, and Sophia.
 - Music: Some custom tracks added.
 
 ## v49 (2021-10-11)